py/PlatformNine/DependaBotTest/sample_django_app/pyproject.toml
[project]
name = "sample_django_app"
version = "0.1.0"
description = "Sample Django application with user management"
requires-python = ">=3.13"
dependencies = [
    "django>=5.0.0",
    "psycopg2-binary>=2.9.9",
    "bcrypt>=4.1.2",
    "djangorestframework>=3.14.0",
    "django-filter>=23.5",
    "gunicorn>=21.2.0",
    "pymemcache>=4.0.0",
    "yarl>=1.9.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["core", "users"] 